+++ priv/components/radio_card.exs
@@ -1,0 +1,32 @@
+[
+  radio_card: [
+    name: "radio_card",
+    args: [
+      variant: ["default", "outline", "shadow", "bordered", "base"],
+      color: [
+        "base",
+        "natural",
+        "primary",
+        "secondary",
+        "success",
+        "warning",
+        "danger",
+        "info",
+        "silver",
+        "misc",
+        "dark",
+        "white",
+        "dawn"
+      ],
+      size: ["extra_small", "small", "medium", "large", "extra_large"],
+      space: ["extra_small", "small", "medium", "large", "extra_large", "none"],
+      padding: ["extra_small", "small", "medium", "large", "extra_large"],
+      rounded: ["extra_small", "small", "medium", "large", "extra_large"],
+      only: ["radio_card"],
+      helpers: [],
+      module: ""
+    ],
+    optional: [],
+    necessary: []
+  ]
+]